requirements.txt
langchain-openai
openai
langchain
langchain_community
langchain_core
python-dotenv
pypdf
arxiv
PyMuPDF
lxml
langchain_huggingface
langchain-text-splitters
faiss-cpu
langchain_chroma
chromadb
langchain_groq
streamlit
fastapi
langserve
uvicorn
bs4
ipykernel
wikipedia
sentence_transformers
duckdb
duckduckgo-search
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
tiktoken
datasets
cassio
PyPDF2
gradio
langchain_nvidia_ai_endpoints
crewai
crewai_tools
nltk
neo4j==5.14
lamini
langchain_experimental